requirements.txt
streamlit 
PyMuPDF 
langchain 
openai 
python-dotenv
PyPDF2
python-docx